modularize electron main process code
We improve code quality in `native/index.ts`

* Extract `WindowManager` to its own module to reduce module size
* Move code running in module scope into the `main` function.
